refactor: scrollToTop 구현에 따라 기존 코드 리펙터링
